requirements.txt
PyPDF2==3.1.0
python-dotenv==1.0.0
google-adk==0.1.0  # Replace with actual ADK package version if different

